Products model for new product creation

+++ controllers/productsController.js
@@ -3,6 +3,9 @@ const fs = require('fs');
 const productsFilePath = path.join(__dirname, '../data/products.json');
 const products = JSON.parse(fs.readFileSync(productsFilePath, 'utf-8'));
 
+
+const Product = require('../models/Product');
+
 // ************ Express Validator Require ************
 const { validationResult } = require('express-validator');
 
@@ -18,35 +21,37 @@ const productsControlador = {
         res.render('./products/addProduct');
     },
     store: (req, res) => {
-        let errors = validationResult(req);
-		const product = {
-			id: Date.now(),
-			brand: req.body.brand,
-            flavor: req.body.flavor,
-			price1: parseInt(req.body.price1),
-            price2: parseInt(req.body.price2),
-            price3: parseInt(req.body.price3),
-            present1: "1/2 litro",
+        const resultValidation = validationResult(req);
+
+        if (!resultValidation.isEmpty()) {
+			return res.render('./products/addProduct', { 
+				errors : resultValidation.mapped(), 
+				oldData : req.body 
+			});
+		}
+
+        let productToCreate = {
+			...req.body,
+			present1: "1/2 litro",
             present2: "1 litro",
             present3: "1.89 litro",
+            price1: parseInt(req.body.price1),
+            price2: parseInt(req.body.price2),
+            price3: parseInt(req.body.price3),
             cant1: parseInt(req.body.cant1),
             cant2: parseInt(req.body.cant2),
             cant3: parseInt(req.body.cant3),
-			description: req.body.des 	
 		}
+
         if (req.file === undefined) {
-            product.image = 'default-image.png';
+            productToCreate.image = 'default-image.png';
           } else {
-            product.image = req.file.filename;
-          }
-         if (errors.isEmpty()) {
-		products.push(product); 
-		productsJSON = JSON.stringify(products, null, 2);
-		fs.writeFileSync(productsFilePath, productsJSON);
-		res.redirect('/products/productList');
-        } else {  
-            res.render('./products/addProduct', { errors: errors.mapped(), old: req.body });
+            productToCreate.image = req.file.filename;
         }
+
+		let productCreated = Product.create(productToCreate);
+
+		return res.redirect('/products/productList');
     },
     edit: (req,res)=>{
         const id = parseInt(req.params.id);
@@ -88,9 +93,10 @@ const productsControlador = {
 		const productToDelete = products.find(product => product.id === id);
 
 
-        fs.unlinkSync('./public/img/products/' + productToDelete.image);
-
-		
+        if (productToDelete.image != 'default-image.png') {
+            fs.unlinkSync('./public/img/products/' + productToDelete.image);
+        }
+        		
 		products.splice(products.indexOf(productToDelete), 1);
 		
 		productsJSON = JSON.stringify(products, null, 2);
